requirements.txt
attrs==23.2.0
blinker==1.7.0
branca==0.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
fiona==1.9.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
folium==0.16.0
GDAL==3.8.5
GeoAlchemy2==0.14.7
geopandas==0.14.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
psycopg2-binary==2.9.9
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
shapely==2.0.4
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
xyzservices==2024.4.0
